refactor(map): Rename map into TreesMap
