+++ src/components/TreesMap/colors.ts
